authentication/forms.py
from django import forms
from django.contrib.auth.forms import UserCreationForm
from .models import CustomUser, UserProfile
from crispy_forms.helper import FormHelper
from crispy_forms.layout import Layout, Submit, Row, Column
class CustomUserCreationForm(UserCreationForm):
    class Meta(UserCreationForm.Meta):
        model = CustomUser
        fields = ('username', 'email', 'password1', 'password2')

class UserProfileForm(forms.ModelForm):
    class Meta:
        model = UserProfile
        fields = ('phone', 'address', 'profile_picture')

    def __init__(self, *args, **kwargs):
            super().__init__(*args, **kwargs)
            self.helper = FormHelper()
            self.helper.layout = Layout(
                Row(
                    Column('phone', css_class='form-group col-md-6'),
                    Column('address', css_class='form-group col-md-6'),
                    css_class='form-row'
                ),
                'profile_picture',
                Submit('submit', 'Update Profile', css_class='btn btn-primary mt-3')
            )

class MFAVerificationForm(forms.Form):
    token = forms.CharField(max_length=6, min_length=6)

def __init__(self, *args, **kwargs):
    super().__init__(*args, **kwargs)
    for field in self.fields.values():
        field.widget.attrs['class'] = 'form-control'