requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap5==2024.10
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.1
django-axes==7.0.1
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-crispy-forms==2.3
django-filter==24.3
django-otp==1.5.4
django-phonenumbers==1.0.1
django-recaptcha==4.0.0
django-storages==1.14.4
django-tinymce==4.1.0
django_csp==3.8
gunicorn==23.0.0
idna==3.10
packaging==24.2
pdf2image==1.17.0
phonenumbers==8.13.52
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-magic==0.4.27
qrcode==8.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
